New folder/tryagain3/requirements.txt
Django==4.2.2
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==8.1.0
pytesseract==0.3.10
